_project-fullstack/crawler/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
playwright>=1.44.0
openai>=1.30.0
qdrant-client>=1.9.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
